+++ Assembly-CSharp/Global/BGSCENE_DEF.cs
@@ -69,8 +69,8 @@ public class BGSCENE_DEF
     public Int16 minY;
     public Int16 maxY;
 
-    public Int16 scrX;
-    public Int16 scrY;
+    public float scrX;
+    public float scrY;
 
     public String name;
